fix: Corregir y mejorar controladores de autenticación y tareas
Mejoras en controladores existentes:
- Corrige problema en task.controller.js que contenía código duplicado
- Mejora documentación JSDoc en ambos controladores
- Implementa manejo de cookies para refreshToken en autenticación
- Añade endpoint de logout
- Mejora validación y normalización de datos
